+++ src/main/java/com/pingguo/bloomtest/BloomtestApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@MapperScan("com.pingguo.bloomtest.dao")
+@MapperScan("com.pingguo.bloomtest.mapper")
 public class BloomtestApplication {
 
     public static void main(String[] args) {
