+++ src/main/java/com/pingguo/bloomtest/controller/ApiDefinitionController.java
@@ -4,7 +4,7 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.pingguo.bloomtest.common.Result;
 import com.pingguo.bloomtest.controller.request.ApiRunTestRequest;
 import com.pingguo.bloomtest.pojo.ApiDefinition;
-import com.pingguo.bloomtest.service.ApiDefinitionService;
+import com.pingguo.bloomtest.service.impl.ApiDefinitionServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,21 +14,21 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("apiDefinition")
 public class ApiDefinitionController {
     @Autowired
-    ApiDefinitionService apiDefinitionService;
+    ApiDefinitionServiceImpl apiDefinitionServiceImpl;
 
     @GetMapping("/list/{projectId}/{moduleId}/{currentPage}/{pageSize}")
     public Result list(@PathVariable Long projectId,
                        @PathVariable Long moduleId,
                        @PathVariable int currentPage,
                        @PathVariable int pageSize) {
-        IPage<ApiDefinition> IPageProject = apiDefinitionService.list(projectId, moduleId, currentPage, pageSize);
+        IPage<ApiDefinition> IPageProject = apiDefinitionServiceImpl.list(projectId, moduleId, currentPage, pageSize);
         return Result.success(IPageProject);
     }
 
     @PostMapping("/add")
     public Result add(@RequestBody ApiDefinition request) {
         try {
-            apiDefinitionService.add(request);
+            apiDefinitionServiceImpl.add(request);
             return Result.success();
         } catch (Exception e) {
             return Result.fail(e.toString());
@@ -37,18 +37,18 @@ public class ApiDefinitionController {
 
     @GetMapping("/getApi")
     public Result getApiById(Long id) {
-        return Result.success(apiDefinitionService.getApi(id));
+        return Result.success(apiDefinitionServiceImpl.getApi(id));
     }
 
     @PostMapping("/postApi")
     public Result postApiById(Long id) {
-        return Result.success(apiDefinitionService.getApi(id));
+        return Result.success(apiDefinitionServiceImpl.getApi(id));
     }
 
     @PostMapping("/update")
     public Result update(@RequestBody ApiDefinition request) {
         try {
-            apiDefinitionService.update(request);
+            apiDefinitionServiceImpl.update(request);
             return Result.success();
         } catch (Exception e) {
             return Result.fail(e.toString());
@@ -57,6 +57,6 @@ public class ApiDefinitionController {
 
     @PostMapping("/apiTestRun")
     public Result apiTestRun(@RequestBody ApiRunTestRequest apiRunTestRequest) {
-        return Result.success(apiDefinitionService.apiTestRun(apiRunTestRequest));
+        return Result.success(apiDefinitionServiceImpl.apiTestRun(apiRunTestRequest));
     }
 }
