+++ src/main/java/com/pingguo/bloomtest/controller/ApiModuleController.java
@@ -3,7 +3,7 @@ package com.pingguo.bloomtest.controller;
 import com.pingguo.bloomtest.common.Result;
 import com.pingguo.bloomtest.controller.request.EditNodeRequest;
 import com.pingguo.bloomtest.pojo.ApiModule;
-import com.pingguo.bloomtest.service.ApiModuleService;
+import com.pingguo.bloomtest.service.impl.ApiModuleServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,18 +12,18 @@ import org.springframework.web.bind.annotation.*;
 public class ApiModuleController {
 
     @Autowired
-    ApiModuleService apiModuleService;
+    ApiModuleServiceImpl apiModuleServiceImpl;
 
     @GetMapping("/list/{projectId}")
     public Result getNodeByProjectId(@PathVariable Long projectId) {
-        return Result.success(apiModuleService.getNodeTreeByProjectId(projectId));
+        return Result.success(apiModuleServiceImpl.getNodeTreeByProjectId(projectId));
     }
 
     @PostMapping("/add")
     public Result addNode(@RequestBody ApiModule node) {
         try {
             System.out.println(node);
-            Long nodeId = apiModuleService.addNode(node);
+            Long nodeId = apiModuleServiceImpl.addNode(node);
             return Result.success(nodeId);
         } catch (Exception e) {
             return Result.fail(e.toString());
@@ -33,7 +33,7 @@ public class ApiModuleController {
     @PostMapping("/rename")
     public Result rename(@RequestBody EditNodeRequest request) {
         try {
-            apiModuleService.renameNode(request);
+            apiModuleServiceImpl.renameNode(request);
             return Result.success();
         } catch (Exception e) {
             return Result.fail(e.toString());
@@ -43,7 +43,7 @@ public class ApiModuleController {
     @GetMapping("/delete/{id}")
     public Result deleteNode(@PathVariable Long id) {
         try {
-            int result = apiModuleService.deleteNode(id);
+            int result = apiModuleServiceImpl.deleteNode(id);
             return Result.success(result);
         } catch (Exception e) {
             return Result.fail(e.toString());
