+++ src/main/java/com/pingguo/bloomtest/controller/ProjectController.java
@@ -3,7 +3,7 @@ package com.pingguo.bloomtest.controller;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.pingguo.bloomtest.common.Result;
 import com.pingguo.bloomtest.pojo.Project;
-import com.pingguo.bloomtest.service.ProjectService;
+import com.pingguo.bloomtest.service.impl.ProjectServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,12 +12,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("project")
 public class ProjectController {
     @Autowired
-    ProjectService projectService;
+    ProjectServiceImpl projectServiceImpl;
 
     @PostMapping("/add")
     public Result addProject(@RequestBody Project project) {
         try {
-            projectService.addProject(project);
+            projectServiceImpl.addProject(project);
             return Result.success();
         } catch (Exception e) {
             return Result.fail(e.toString());
@@ -27,7 +27,7 @@ public class ProjectController {
     @GetMapping("/getAdd")
     public Result getAddProject(@RequestBody Project project) {
         try {
-            projectService.addProject(project);
+            projectServiceImpl.addProject(project);
             return Result.success();
         } catch (Exception e) {
             return Result.fail(e.toString());
@@ -38,20 +38,20 @@ public class ProjectController {
     public Result getProjectList(@PathVariable int currentPage,
                                  @PathVariable int pageSize,
                                  @RequestBody Project project) {
-        IPage<Project> IPageProject = projectService.getProjectList(currentPage, pageSize, project);
+        IPage<Project> IPageProject = projectServiceImpl.getProjectList(currentPage, pageSize, project);
         return Result.success(IPageProject);
 
     }
 
     @GetMapping("/list/all")
     public Result getProjectAll() {
-        return Result.success(projectService.getProjectAll());
+        return Result.success(projectServiceImpl.getProjectAll());
     }
 
     @PostMapping("/update")
     public Result updateProject(@RequestBody Project project) {
         try {
-            projectService.updateProject(project);
+            projectServiceImpl.updateProject(project);
             return Result.success();
         } catch (Exception e) {
             return Result.fail(e.toString());
