+++ src/main/java/com/pingguo/bloomtest/mapper/ApiDefinitionMapper.java
@@ -1,4 +1,4 @@
-package com.pingguo.bloomtest.dao;
+package com.pingguo.bloomtest.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.pingguo.bloomtest.pojo.ApiDefinition;
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ApiDefinitionDAO extends BaseMapper<ApiDefinition> {
+public interface ApiDefinitionMapper extends BaseMapper<ApiDefinition> {
 }
