+++ src/main/java/com/pingguo/bloomtest/mapper/ApiModuleMapper.java
@@ -1,13 +1,10 @@
-package com.pingguo.bloomtest.dao;
+package com.pingguo.bloomtest.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.pingguo.bloomtest.pojo.ApiModule;
-import org.apache.ibatis.annotations.Select;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 
 @Repository
-public interface ApiModuleDAO extends BaseMapper<ApiModule> {
+public interface ApiModuleMapper extends BaseMapper<ApiModule> {
 }
