+++ src/main/java/com/pingguo/bloomtest/mapper/ProjectMapper.java
@@ -1,9 +1,9 @@
-package com.pingguo.bloomtest.dao;
+package com.pingguo.bloomtest.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.pingguo.bloomtest.pojo.Project;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectDAO extends BaseMapper<Project> {
+public interface ProjectMapper extends BaseMapper<Project> {
 }
