+++ src/main/java/com/pingguo/bloomtest/mapper/UserDemoMapperPlus.java
@@ -1,7 +1,7 @@
-package com.pingguo.bloomtest.dao;
+package com.pingguo.bloomtest.mapper;
 
 import com.pingguo.bloomtest.pojo.User;
 
-public interface UserMapperPlus {
+public interface UserDemoMapperPlus {
     User getUserById(Integer id);
 }
