+++ src/main/java/com/pingguo/bloomtest/mapper/UserMapper.java
@@ -1,10 +1,13 @@
-package com.pingguo.bloomtest.dao;
+package com.pingguo.bloomtest.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.pingguo.bloomtest.pojo.User;
 import org.springframework.stereotype.Repository;
 
+/**
+ * @author weikang.sang
+ * @date 23:36 2022/5/10
+ */
 @Repository
-public interface UserDAO extends BaseMapper<User> {
-
+public interface UserMapper extends BaseMapper<User> {
 }
