+++ src/main/resources/UserMapperPlus.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.pingguo.bloomtest.dao.UserMapperPlus">
+<mapper namespace="com.pingguo.bloomtest.mapper.UserDemoMapperPlus">
     <select id="getUserById" resultType="com.pingguo.bloomtest.pojo.User">
         select * from user where id = #{id}
     </select>
