+++ src/main/resources/application.properties
@@ -1,9 +1,9 @@
 #server.port=5544
 
 # mysql
-spring.datasource.url=jdbc:mysql://xxxx:3306/bloomtest?characterEncoding=UTF-8&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://119.3.84.101:3306/bloomtest?characterEncoding=UTF-8&serverTimezone=UTC
 spring.datasource.username=root
-spring.datasource.password=xxxx
+spring.datasource.password=Sang@wk080721
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 #此属性控制客户端（即您）等待来自池的连接的最大毫秒数
@@ -28,4 +28,3 @@ mybatis-plus.configuration.map-underscore-to-camel-case=false
 #mybatis-plus.configuration.log-impl=org.apache.ibatis.logging.stdout.StdOutImpl
 
 spring.jackson.date-format=yyyy-MM-dd HH:mm:ss
-spring.jackson.time-zone=GMT+8
